requirements.txt
# ============================================================================
# AI TEACHING ASSISTANT PRO - COMPLETE REQUIREMENTS
# Python 3.8+ | Virtual Environment Recommended
# ============================================================================

# FRONTEND
# ============================================================================
streamlit>=1.28.0                    # Web UI framework
plotly>=5.15.0                       # Interactive charts
pandas>=2.0.0                        # Data manipulation
numpy>=1.24.0                        # Numerical computing

# LLM AND RAG
# ============================================================================
langchain>=0.1.0                     # LLM framework
langchain-community>=0.0.20          # LangChain community extensions
langchain-core>=0.1.0                # Core LangChain
openai>=1.3.0                        # OpenAI API

# VECTOR DATABASE & EMBEDDINGS
# ============================================================================
faiss-cpu>=1.7.4                     # Vector similarity search
sentence-transformers>=2.2.2         # Text embeddings

# DATA PROCESSING
# ============================================================================
pypdf>=3.17.0                        # PDF reading
beautifulsoup4>=4.12.0               # HTML parsing
requests>=2.31.0                     # HTTP requests

# MACHINE LEARNING & DATA SCIENCE
# ============================================================================
scikit-learn>=1.3.0                  # ML algorithms
nltk>=3.8.0                          # NLP toolkit
matplotlib>=3.7.0                    # Plotting
seaborn>=0.12.0                      # Statistical plots

# DATABASE
# ============================================================================
# SQLite3 (included in Python)

# UTILITIES
# ============================================================================
python-dotenv>=1.0.0                 # Environment variables

# ============================================================================
# INSTALLATION:
# 1. Create venv: python -m venv venv
# 2. Activate:    venv\Scripts\activate  (Windows)
#                 source venv/bin/activate  (Mac/Linux)
# 3. Install:     pip install -r requirements.txt
# ============================================================================